build: upgrade ErrorProne to 2.37.0 and NullAway to 0.13.7
Fix all new warnings introduced by ErrorProne 2.37.0:
- StatementSwitchToExpressionSwitch: convert statement switches to
  expression switches across config, consensus, evm, ethereum modules
- AssignmentExpression: refactor compound assignments in SarOperation,
  ShlOperation, ShrOperation, StreamingOperationTracer, Era1Reader,
  AbstractDebugOperationTracer
- ReferenceEquality: suppress in BytesValueRLPOutput, EVM,
  TransactionSelectionResult where identity equality is intentional
- CollectionUndefinedEquality: suppress in EthScheduler and
  AbstractEthTask for CompletableFuture collections
- InlineFormatString: fix in Memory.java
- EffectivelyPrivate: make JsonUtil constructor package-private
- NotJavadoc: fix in UpnpNatManager

Update dependency verification metadata for new artifact checksums.

+++ evm/src/main/java/org/hyperledger/besu/evm/frame/MessageFrame.java
@@ -791,7 +791,7 @@ public class MessageFrame {
       final int srcSize = value.size();
       if (length > srcSize) {
         final MutableBytes paddedAnswer = MutableBytes.create((int) length);
-        if ((long) 0 < srcSize) {
+        if (0L < srcSize) {
           value.slice(0, srcSize).copyTo(paddedAnswer, (int) (length - srcSize));
         }
         setUpdatedMemory(offset, paddedAnswer.copy());
